Update ItemSalary : - Change type Attibute + LimitMny from Long to BigDecimal

+++ nts.uk/ctx/pr/core/nts.uk.ctx.pr.core.app/src/main/java/nts/uk/ctx/pr/core/app/command/itemmaster/itemsalary/AddItemSalaryCommand.java
@@ -30,5 +30,5 @@ public class AddItemSalaryCommand {
 	private String memo;
 	private int limitMnyAtr;
 	private String limitMnyRefItemCd;
-	private Long limitMny;
+	private BigDecimal limitMny;
 }
